Add tests for chromatic adaptation methods
I sort of check my methods here with my own results which is somehow awkward...
Since I found no sample data in the Internet this was the only solution I found.
However I compared the resulting colors with the input colors and decided if the result seems to be ok.